# project-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Actor Platform API",
    "version": "0.0.1",
    "description": "API for managing actors, runs, schedules, storage, and webhooks on the actor automation platform."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "paths": {
...
    "/workspaces/{workspaceId}/webhooks": {
      "get": {
        "summary": "List webhooks",
        "tags": ["Webhooks"],
        "security": [{ "bearerAuth": ["webhooks:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated webhook list", "content": { "application/json": { "schema": { "type": "object", "properties": { "webhooks": { "type": "array", "items": { "$ref": "#/components/schemas/Webhook" } }, "nextCursor": { "type": "string" } } } } } } }
      },
      "post": {
        "summary": "Create a webhook",
        "tags": ["Webhooks"],
        "security": [{ "bearerAuth": ["webhooks:write"] }],
        "parameters": [{ "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWebhookInput" } } } },
        "responses": { "201": { "description": "Webhook created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } } }
      }
    },
    "/workspaces/{workspaceId}/webhooks/{webhookId}": {
      "get": {
        "summary": "Get webhook by ID",
        "tags": ["Webhooks"],
        "security": [{ "bearerAuth": ["webhooks:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "webhookId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Webhook details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } } }
      },
      "patch": {
        "summary": "Update a webhook",
        "tags": ["Webhooks"],
        "security": [{ "bearerAuth": ["webhooks:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "webhookId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateWebhookInput" } } } },
        "responses": { "200": { "description": "Webhook updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } } }
      },
      "delete": {
        "summary": "Delete a webhook",
        "tags": ["Webhooks"],
        "security": [{ "bearerAuth": ["webhooks:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "webhookId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Webhook deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/workspaces/{workspaceId}/webhooks/{webhookId}/attempts": {
      "get": {
        "summary": "List webhook delivery attempts",
        "tags": ["Webhooks"],
        "security": [{ "bearerAuth": ["webhooks:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "webhookId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated attempt list", "content": { "application/json": { "schema": { "type": "object", "properties": { "attempts": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookAttempt" } }, "nextCursor": { "type": "string" } } } } } } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SHA-256 hashed API token"
      }
    },
    "schemas": {
...
      "SuccessResponse": { "type": "object", "properties": { "success": { "type": "boolean" } }, "required": ["success"] },
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "workspaceId": { "type": "string" }, "actorId": { "type": "string" },
          "ownerId": { "type": "string" }, "eventTypes": { "type": "string" }, "url": { "type": "string", "format": "uri" },
          "secret": { "type": "string" }, "enabled": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "WebhookAttempt": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "webhookId": { "type": "string" }, "eventType": { "type": "string" },
          "status": { "type": "string", "enum": ["PENDING", "DELIVERING", "RETRYING", "SUCCEEDED", "FAILED"] },
          "attemptCount": { "type": "integer" }, "requestUrl": { "type": "string" },
          "requestBody": { "type": "string" }, "responseStatusCode": { "type": "integer" },
          "responseBody": { "type": "string" }, "errorMessage": { "type": "string" },
          "nextRetryAt": { "type": "string", "format": "date-time" },
          "lastRetryAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }, "completedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateWebhookInput": {
        "type": "object",
        "required": ["actorId", "eventTypes", "url"],
        "properties": {
          "actorId": { "type": "string" }, "eventTypes": { "type": "string" },
          "url": { "type": "string", "format": "uri" }, "secret": { "type": "string" },
          "enabled": { "type": "boolean" }
        }
      },
      "UpdateWebhookInput": {
        "type": "object",
        "properties": {
          "eventTypes": { "type": "string" }, "url": { "type": "string", "format": "uri" },
          "secret": { "type": "string" }, "enabled": { "type": "boolean" }
        }
      }
    }
  }
}